[ex/bert_training] Slight refactor of the example
